Raise-warnings-if-value/pyproject.toml
[tool.poetry]
name = "raise-warnings-if-value"
version = "0.1.0"
description = ""
authors = ["Saikat Karmakar <you@example.com>"]
readme = "README.md"
packages = [{include = "raise_warnings_if_value"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
#eth-ape = "^0.6.15"
#ape-foundry = "^0.6.14"
eth-ape = {path = "/home/avik/Desktop/git_projects/contributions/ape"}
ape-foundry = "^0.6.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
